commands/blscfg: discover UKI on EFI
Add a new command "blscfg_uki" to the blscfg module.  It looks for
unified kernel images (UKI) in the locations defined by the Boot Loader
Specification.  Additionally check on the device indicated by the $root
variable.  Add all UKIs found to the menu.

The UKI are booted using the 'chainloader' command on EFI.

Signed-off-by: Oliver Steffen <osteffen@redhat.com>