commands/bli: Extract uft8 to utf16 conversion
Create a new function for utf-8 to utf-16 conversion called
`grub_utf8_to_utf16_alloc` in the bli module.  It is modeled after the
`grub_utf8_to_ucs4_alloc` like functions in charset.h.  The goal is to
move this to a central place so that it can help reduce code
duplication.

Please comment: Where should this go?